+++ node/src/types/kernel/api/template/renderSprig/response.d.ts
@@ -15,15 +15,20 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import "dotenv/config";
-import { Client } from "@/client/Client";
-
-const client = new Client({
-    baseURL: process.env.VITE_SIYUAN_SERVE,
-    token: process.env.VITE_SIYUAN_TOKEN,
-});
-
-export default {
-    Client,
-    client,
-};
+/**
+ * render Sprig template
+ */
+export interface IResponse {
+    /**
+     * status code
+     */
+    code: number;
+    /**
+     * the result of rendering the Sprig template
+     */
+    data: string;
+    /**
+     * status message
+     */
+    msg: string;
+}