+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # 更改日志 | Change Log
 
+- 添加 API `/api/template/renderSprig` | Add API `/api/template/renderSprig`
+
 ## v0.0.3 / 2023-06-29
 
 - [v0.0.2 ... v0.0.3](https://github.com/siyuan-community/siyuan-sdk/compare/v0.0.2...v0.0.3)